customized list model
