+++ internal/cli/views/basics/list/item.go
@@ -1,4 +1,4 @@
-package item
+package list
 
 import (
 	"fmt"
